% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.translator.R
\name{rpart.translator}
\alias{rpart.translator}
\title{RPart Translator}
\usage{
rpart.translator(model, data)
}
\arguments{
\item{model}{The model generated by "rpart" containing the decision tree.}

\item{data}{The data that was used to create the "rpart" model.}
}
\value{
A data table of rules and sub-rules. Each column is listed below:
\item{Rule}{ Identifier for an overall rule that many sub-rules roll up
into. A rule requires all sub-rules to be applied to a data set. }
\item{Subrule}{ Identifier for a single piece of a larger rule. There can be
many sub-rules that make up a single rule. }
\item{Variable}{ Column name of
the data set that the sub-rule is defining. }
\item{Operator}{ Mathematical
operator defining how a column and value interact. Can be =, >=, <=, >, or
<. A sub-rule with many "=" operators are treated as "OR" staements. }
\item{Value}{ Value of the column (defined in the Variable field) that the
sub-rule is defining. }
\item{TotalVolume}{ Number of records that the sub-rule
captures when all sub-rules are applied for a single rule. }
\item{TargetPredicted}{ The target value the sub-rule is trying to capture. }
\item{TargetPredictedVolume}{ Number of target records that the sub-rule
captures when all sub-rules are applied for a single rule. }
}
\description{
This function formats a "rpart" decision tree model into a table of rules
and sub-rules.
}
\details{
Before running this function, a "rpart" model must be generated using the
"rpart" package. That model and the data used to generate the model will be
passed in as parameters to create the table of rules and sub-rules, with
accompanying volume statistics.
}
\examples{

library(rpart)
library(rpart.translator)

fit <- rpart(Kyphosis ~ Age + Number + Start, data = kyphosis)

dataTable <- rpart.translator(fit, kyphosis)

}
\references{
Terry Therneau and Beth Atkinson (2018) Authors and
maintainers of rpart version 4.1-13. Brian Ripley (1999-2017) Producer of
the initial R port and maintainer.
}
\keyword{~kwd1}
\keyword{~kwd2}
